+++ src/uu/printf/Cargo.toml
@@ -20,7 +20,7 @@ path = "src/printf.rs"
 [dependencies]
 clap = { version = "2.33", features = ["wrap_help"] }
 itertools = "0.8.0"
-uucore = { version=">=0.0.11", package="uucore", path="../../uucore" }
+uucore = { version=">=0.0.11", package="uucore", path="../../uucore", features=["memo"] }
 uucore_procs = { version=">=0.0.8", package="uucore_procs", path="../../uucore_procs" }
 
 [[bin]]
