+++ src/uu/printf/src/mod.rs
@@ -1,3 +1,1 @@
 mod cli;
-mod memo;
-mod tokenize;
