+++ src/uu/printf/src/printf.rs
@@ -4,10 +4,8 @@
 
 use clap::{crate_version, App, Arg};
 use uucore::error::{UResult, UUsageError};
+use uucore::memo;
 use uucore::InvalidEncodingHandling;
-
-mod memo;
-mod tokenize;
 
 const VERSION: &str = "version";
 const HELP: &str = "help";
